+++ service/controller/resource/prometheus/resource.go
@@ -163,8 +163,14 @@ func toPrometheus(v interface{}, config Config) (metav1.Object, error) {
 			Replicas: &replicas,
 			Resources: corev1.ResourceRequirements{
 				Requests: corev1.ResourceList{
-					// cpu: 100m
-					corev1.ResourceCPU: *resource.NewMilliQuantity(100, resource.DecimalSI),
+					// cpu: 1000m
+					corev1.ResourceCPU: *resource.NewMilliQuantity(1000, resource.DecimalSI),
+					// memory: 5Gi
+					corev1.ResourceMemory: *resource.NewQuantity(5000*1024*1024, resource.BinarySI),
+				},
+				Limits: corev1.ResourceList{
+					// cpu: 1000m
+					corev1.ResourceCPU: *resource.NewMilliQuantity(1000, resource.DecimalSI),
 					// memory: 5Gi
 					corev1.ResourceMemory: *resource.NewQuantity(5000*1024*1024, resource.BinarySI),
 				},