+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Set prometheus cpu requests and limits to 1 CPU.
+- Set prometheus memory requests and limits to 5Gi.
+
 ## [1.10.0] - 2020-11-20
 
 ### Added